Add files via upload
The WebScraper application was used to prepare dataset according to our needs. For instance our dataset consisted of various celebrity looks which have been popular over internet and social media. The webscraper uses BeautifulSoup package for this purpose. 

Look_Recreation is the ML model where images are processed and the model is trained. We have taken four categories for the data in this file, but it can be customised according to data availability. 

The app.py is flask application which is linked to index.html to provide a web interface, where users can upload a look they want to search for and get the products to buy the look.